LayoutTests/http/tests/security/contentSecurityPolicy/object-src-param-src-blocked2-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/object-src-param-src-blocked2.html because it does not appear in the object-src directive of the Content Security Policy.
